+++ Kl.Tests/ConditionalEvaluation.fs
@@ -1,7 +1,6 @@
 ﻿module Kl.Tests.``Conditional Evaluation``
 
 open NUnit.Framework
-open Kl
 open Kl.Values
 open Kl.Startup
 open Assertions