+++ Kl.Tests/EvaluationOrder.fs
@@ -2,8 +2,6 @@
 
 open NUnit.Framework
 open Kl
-open Kl.Values
-open Kl.Startup
 open Assertions
 
 [<Test>]