+++ Kl.Tests/PartialApplication.fs
@@ -3,7 +3,6 @@
 open NUnit.Framework
 open Kl
 open Kl.Values
-open Kl.Startup
 open Assertions
 
 [<Test>]