+++ Kl.Tests/Properties/Metadata.fs
@@ -1,7 +1,0 @@
-module ShenSharp.Metadata
-
-open System.Reflection
-
-[<assembly: AssemblyTitle "Kl.Tests">]
-
-do ()