+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ## [Unreleased]
 
 **Updated to Shen Open Source Kernel 22.2**
+**Ported to .Net Standard 2.1**
 
 ## [0.9.6.0] - 2018-10-08